fix importing issue

+++ yerbamate/mate.py
@@ -1,14 +1,54 @@
 import os
 from argparse import ArgumentParser
+import importlib
 from pytorch_lightning import Trainer
 from pytorch_lightning.callbacks import EarlyStopping
 from types import SimpleNamespace
 from torch import nn
 import torch as t
+import ipdb
 import json
-
+import sys
 # TODO: find root recursively
-# TODO: 
+# TODO:
+
+import imp
+import sys
+
+
+# dynamic import
+def dynamic_imp(name, class_name):
+
+    # find_module() method is used
+    # to find the module and return
+    # its description and path
+    try:
+
+        ipdb.set_trace()
+        fp, path, desc = imp.find_module(name)
+
+    except ImportError:
+        print("module not found: " + name)
+
+    try:
+        # load_modules loads the module
+        # dynamically ans takes the filepath
+        # module and description as parameter
+        example_package = imp.load_module(name, fp, path, desc)
+
+    except Exception as e:
+        print(e)
+
+    try:
+        myclass = imp.load_module(
+            "% s.% s" % (name, class_name), fp, path, desc
+        )
+
+    except Exception as e:
+        print(e)
+
+    return example_package, myclass
+
 
 class Mate:
     def __init__(self):
@@ -16,7 +56,7 @@ class Mate:
         self.save_path = ""
         self.current_folder = os.path.dirname(__file__)
         self.__findroot()
-        self.models = self.__list_packages('models')
+        self.models = self.__list_packages("models")
 
     def __list_packages(self, folder: str):
         return (
@@ -35,46 +75,61 @@ class Mate:
         current_path = os.getcwd()
         self.root_folder = os.path.basename(current_path)
         os.chdir("..")
+        sys.path += '.'
 
     def __load_model_class(self, model_name: str, folder="models"):
-        return __import__(
-            f"{self.root_folder}.{folder}.{model_name}.model"
+        return  __import__(
+            f"{self.root_folder}.{folder}.{model_name}.model",
+            fromlist=[folder],
         ).Model
 
     def __load_logger_class(self):
-        return __import__(f"{self.current_folder}.logger")
+        return __import__(f"{self.current_folder}.logger").CustomLogger
 
     def __load_data_loader_class(self, data_loader_name: str):
         return __import__(
-            f"{self.root_folder}.{data_loader_name}.{data_loader_name}"
+            f"{self.root_folder}.data_loader.{data_loader_name}",
+            fromlist=['data_loader']
         ).CustomDataModule
 
     def __set_save_path(self, model_name: str):
         self.save_path = os.path.join(self.root_folder, "models", model_name)
 
-    def __read_parameters(self, model_name:str):
-        with open(os.path.join(self.root_folder, "models", model_name, "parameters.json")) as f:
+    def __read_parameters(self, model_name: str):
+        with open(
+            os.path.join(
+                self.root_folder, "models", model_name, "parameters.json"
+            )
+        ) as f:
             params = json.load(f)
-        self.params = SimpleNamespace(**params)
+        return SimpleNamespace(**params)
 
-    def __get_trainer(self, data_loader_name:str):
-        data_module = self.__load_data_loader_class(data_loader_name)(self.params)
+    def __get_trainer(self, model_name: str):
+        params = self.__read_parameters(model_name)
+        model = self.__load_model_class(model_name)(params)
+        data_module = self.__load_data_loader_class(params.data_loader)(
+            params
+        )
         logger_module = self.__load_logger_class()
-        return Trainer(
-            max_epochs=self.params.max_epochs,
-            gpus=(1 if self.params.cuda else None),
-            callbacks=[
-                EarlyStopping(
-                    monitor="val_loss",
-                    patience=self.params.early_stopping_patience,
-                ),
-                # checkpoint_callback,
-            ],
-            logger=logger_module.CustomLogger(self.params),
-            enable_checkpointing=False,
+        return (
+            Trainer(
+                max_epochs=params.max_epochs,
+                gpus=(1 if params.cuda else None),
+                callbacks=[
+                    EarlyStopping(
+                        monitor="val_loss",
+                        patience=params.early_stopping_patience,
+                    ),
+                    # checkpoint_callback,
+                ],
+                logger=logger_module(params),
+                enable_checkpointing=False,
+            ),
+            model,
+            data_module,
         )
 
-    def __load_model(self, checkpoint_path:str, model: nn.Module):
+    def __load_model(self, checkpoint_path: str, model: nn.Module):
         if os.path.exists(checkpoint_path):
             print(f"\nLoading model from checkpoint:'{checkpoint_path}'\n")
             model.load_state_dict(t.load(checkpoint_path))
@@ -82,7 +137,6 @@ class Mate:
             raise Exception(
                 "No checkpoint found. You must train the model first!"
             )
-
 
     def init(self):
         pass
@@ -104,34 +158,48 @@ class Mate:
         else:
             print("Ok, exiting.")
 
-    def list(self, folder:str):
-        print("\n".join(tuple('\t' + str(m) for m in self.__list_packages(folder))))
+    def list(self, folder: str):
+        print(
+            "\n".join(
+                tuple("\t" + str(m) for m in self.__list_packages(folder))
+            )
+        )
 
     def clone(self, source_model: str, target_model: str):
         os.system(
             f"cp -r {os.path.join(self.root_folder, 'models', source_model)} {os.path.join(self.root_folder, 'models', target_model)}"
         )
 
+    def __fit(self, model_name: str):
+        trainer, model, data_module = self.__get_trainer(model_name)
+        trainer.fit(model, datamodule=data_module)
+
     def train(self, model_name: str):
-        self.__read_parameters(model_name)
+        assert (
+            model_name in self.models
+        ), f'Model "{model_name}" does not exist.'
+
         save_path = os.path.join(self.root_folder, "models", model_name)
         checkpoint_path = os.path.join(save_path, "checkpoint.pt")
         action = "go"
-        while action not in ("y", "n", ""):
-            action = input(
-                "Checkpiont file exists. Re-training will erase it. Continue? ([y]/n)\n"
-            )
-        if action in ("y", ""):
-            os.remove(checkpoint_path)
-        else:
-            print("Ok, exiting.")
-            return
+        if os.path.exists(checkpoint_path):
+            while action not in ("y", "n", ""):
+                action = input(
+                    "Checkpiont file exists. Re-training will erase it. Continue? ([y]/n)\n"
+                )
+            if action in ("y", ""):
+                os.remove(checkpoint_path)
+            else:
+                print("Ok, exiting.")
+                return
+        self.__fit(model_name)
 
     def test(self, model_name: str):
-        self.__read_parameters(model_name)
+        trainer, model, data_module = self.__get_trainer(model_name)
+        trainer.test(model, datamodule=data_module)
 
     def restart(self, model_name: str):
-        self.__read_parameters(model_name)
+        self.__fit(model_name)
 
     def tune(self, model: str, params: tuple[str, ...]):
         pass
